Move sanitised dump daily job to yearly job
Noticed that the execution schedule is controlled by app.json. @monitor
only adjust the sentry monitor. Initially setting the schedule of this
job to yearly, as that isn't run by the dokku cron process configured
in app.json. So it won't run automatically. We can manually run and
verify that the sanitised dump works in this way. Thanks to @mikerkelly
for this suggestion.

Changes covered in this commit:
- Deleted the existing daily dump_sanitised_db.py script and
  allowlist.json and replaced it with a new yearly version.

- Updated the job class to inherit from YearlyJob instead of DailyJob.

- Adjusted the cron schedule for the job to run annually.

- Modified integration and unit tests to reference the new yearly job.

- Fixed 2 unit tests for _fake_expression as they were still expecting
  random values rather than ROW_NUMBER
